# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Player Management Service",
    "description": "API for managing players and teams",
    "version": "1.0.0",
    "license": {
      "name": "Creative Commons Zero v1.0 Universal",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
  "paths": {
...
    "/download/json": {
      "get": {
        "summary": "Download players data in JSON format",
        "description": "Provides the list of players in JSON file format",
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Text to search for in players data",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attribute",
            "in": "query",
            "description": "Attribute to filter by in players data",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON file containing players data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter"
          }
        }
      }
    },
    "/download/csv": {
      "get": {
        "summary": "Download players data in CSV format",
        "description": "Provides the list of players in CSV file format",
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Text to search for in players data",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attribute",
            "in": "query",
            "description": "Attribute to filter by in players data",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file containing players data",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter"
          }
        }
      }
    }
  },
...
}  